requirements.txt
paho_mqtt==2.1.0
python-dotenv==1.0.1
pytest==8.3.3